Resolved crash during FamilyTest.
The cause was an UnitializedFieldError due to some thrashing between the
Family companion object and the Family class (the companion created a
vector of 7 base families, and each family instance tried to examine the
size of that vector before it had been created). Reworked the assertion
and remove the NumBaseFamilies member from the companion.