Undid Scalastyle config file changes.
It appears difficult to configure Scalastyle to find its configuration
file in multi-module SBT builds (the modules still look in the root
folder for a file called "scalastyle-config.xml", regardless of how the
scalastyleConfig property is configured). However, it seems that
IntellIJ IDEA may now support the default scalastyle configuration.
Whether or not that's that case, it's more important that Scalastyle
executes correctly thant whether IntelliJ reuses Scalastyle's config.
correctly.