Renamed problematic "prn" package.
"prn" is a reserved name on Windows (historically, the parallel port
device, PRN:) and cannot be used as a folder name. Consequently,
sources with this package cannot be checkout out using git, as this
results in an error:

fatal: cannot create directory at
'facsimile-stat/src/main/scala/org/facsim/stat/prn': Invalid argument

To overcome this, the package was renamed to be "prng".